6-27

+++ controllers/addNewOrder.js
@@ -20,42 +20,53 @@ module.exports = async (ctx, next)=>{
 	user_id = info.user_id
 	dinning_choice = info.dinning_choice
 
-	//update users
+	//update users(if needed)
 	takeout_info = info.takeout_info
-	queryObj = {}
-	queryObj.table = 'users'
-	queryObj.columns = ['wechat_name','phone','location']
-	queryObj.key = 'user_id'
-	queryObj.keyValue = user_id
-	user_info = await query.query(ctx, next, '', queryObj)
-	wechat_name = !user_info[0].wechat_name?takeout_info.name:user_info[0].wechat_name
-	phone = !user_info[0].phone?takeout_info.phone:user_info[0].phone
-	location = !user_info[0].location?takeout_info.location:user_info[0].location
+	if(takeout_info){
+		console.log('\ntakeoutinfo:',takeout_info,'\n')
+		queryObj = {}
+		queryObj.table = 'users'
+		queryObj.columns = ['wechat_name','phone','location']
+		queryObj.key = 'user_id'
+		queryObj.keyValue = user_id
+		console.log('\nuser_id:',user_id,'\n')
+		user_info = await query.query(ctx, next, '', queryObj)
+		console.log('\nuser_info:',user_info,'\n')
+		wechat_name = !user_info[0].wechat_name?takeout_info.name:user_info[0].wechat_name
+		phone = !user_info[0].phone?takeout_info.phone:user_info[0].phone
+		location = !user_info[0].location?takeout_info.location:user_info[0].location
 
-	sql_update = 'UPDATE `users` SET `wechat_name`=\'' + wechat_name + '\', `location`=\''
-		+ location + '\', `phone`=\'' + phone + '\' WHERE `user_id`=\'' + user_id + '\''
-	await query.query(ctx, next, sql_update, {})
-
-	//get info of discount
-	queryObj={}
-	queryObj.table='coupon'
-	queryObj.columns=['money','amount']
-	queryObj.key='discount_id'
-	queryObj.keyValue=info.discount_id
-	discount_info = await query.query(ctx, next, '', queryObj)
-	discount_value = discount_info[0].money
-	discount_amount = discount_info[0].amount
-	total_price -= discount_value
-
-	//delete discount
-	--discount_amount
-	sql = ''
-	if(discount_amount<=0){
-		sql = 'DELETE FROM `coupon` WHERE `discount_id`=\''+info.discount_id+'\''
-	}else{
-		sql = 'UPDATE `coupon` SET `amount`= \''+discount_amount+'\' WHERE `discount_id`=\''+info.discount_id+'\''
+		sql_update = 'UPDATE `users` SET `wechat_name`=\'' + wechat_name + '\', `location`=\''
+			+ location + '\', `phone`=\'' + phone + '\' WHERE `user_id`=\'' + user_id + '\''
+		await query.query(ctx, next, sql_update, {})
 	}
-	await query.query(ctx, next, sql, {})
+
+	//get info of discount(if needed)
+	if(info.discount_id){
+		queryObj={}
+		queryObj.table='coupon'
+		queryObj.columns=['money','amount']
+		queryObj.key='discount_id'
+		queryObj.keyValue=info.discount_id
+		console.log('\ndiscount_id:',info.discount_id)
+		discount_info = await query.query(ctx, next, '', queryObj)
+		console.log('\ndiscount_info:',discount_info)
+		discount_value = discount_info[0].money
+		discount_amount = discount_info[0].amount
+		total_price -= discount_value
+
+		//delete discount
+		--discount_amount
+		console.log('dscount_amount:',discount_amount)
+		sql = ''
+		if(discount_amount<=0){
+			sql = 'DELETE FROM `coupon` WHERE `discount_id`=\''+info.discount_id+'\''
+		}else{
+			sql = 'UPDATE `coupon` SET `amount`= \''+discount_amount+'\' WHERE `discount_id`=\''+info.discount_id+'\''
+		}
+		console.log('sql_coupon:',sql)
+		await query.query(ctx, next, sql, {})
+	}
 
 	//update orders
 	sql4 = 'INSERT INTO `orders` (`user_id1`, `dinning_choice`,`total_price`,`note`) VALUES (\''
